FSAL_GPFS: build errors on Ubuntu xenial
32-bit string format and used uninitialized

+++ src/FSAL/FSAL_GPFS/handle.c
@@ -799,7 +799,7 @@ static fsal_status_t listxattrs(struct fsal_obj_handle *obj_hdl,
 		next += 1;
 
 		LogDebug(COMPONENT_FSAL,
-		"nameP %s at offset %ld", name, (next - name));
+		"nameP %s at offset %ld", name, ((long)(next - name)));
 
 		if (entryCount >= *la_cookie) {
 			if ((((char *)entry - (char *)lr_names->entries) +
@@ -813,7 +813,7 @@ static fsal_status_t listxattrs(struct fsal_obj_handle *obj_hdl,
 				LogFullDebug(COMPONENT_FSAL,
 				   "out1 cookie %llu off %ld eof %d cookieverf %llx",
 				   (unsigned long long)*la_cookie,
-				   (next - name), *lr_eof,
+				   ((long)(next - name)), *lr_eof,
 				   (unsigned long long)*
 				   ((uint64_t *)la_cookieverf));
